fix: field name rotation on chromium 66
The latest version of Chrome (94 as of this commit) supports the
`writing-mode: vertical-rl` style on table cells (`td` tags) directly.
Cefpython 66.1 uses Chromium 66 which does not support this style on
table cells directly.  Work around by adding a div to the cell and
applying the style on the div instead.

References:

* https://bugs.webkit.org/show_bug.cgi?id=65917
* https://stackoverflow.com/questions/59024097/vertical-text-in-table-cell-using-css3-writing-mode-in-chrome