fix: layout table issues with rotation and inconsistent column widths
A previous fix improved rotation issues on Chromium 66 by applying the
writing-mode style to a div instead of a td.  However, Chromium 66 seems
to have issues with the writing-mode style on a div too.  Replacing the
div with a span makes it even better and removes the need for the
margin styles to center the text.

Additionally, adding 'table-layout: fixed' fixes undesired width changes
to the bit index cells and makes them obey their width styles better.